refactor(dictgen): Pull out custom str
