+++ crates/dictgen/src/lib.rs
@@ -4,6 +4,7 @@
 
 #[cfg(feature = "codegen")]
 mod gen;
+mod insensitive;
 #[cfg(feature = "map")]
 mod map;
 mod table;
@@ -11,6 +12,7 @@ mod trie;
 
 #[cfg(feature = "codegen")]
 pub use gen::*;
+pub use insensitive::*;
 #[cfg(feature = "map")]
 pub use map::*;
 pub use table::*;
